backend/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.3.1
cycler==0.12.1
discord==2.3.2
discord.py==2.3.2
distlib==0.3.8
Django==5.1.1
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
filelock==3.14.0
flake8==7.1.0
fonttools==4.55.3
frozenlist==1.4.1
idna==3.7
kiwisolver==1.4.7
matplotlib==3.9.3
mccabe==0.7.0
multidict==6.0.5
mysql==0.0.3
mysql-connector-python==9.1.0
mysqlclient==2.2.6
numpy==2.2.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
platformdirs==4.2.2
psycopg2-binary==2.9.10
pycodestyle==2.12.0
pyflakes==3.2.0
PyJWT==2.9.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.31.0
six==1.17.0
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.1
virtualenv==20.26.2
yarl==1.9.4
